pyproject.toml
[project]
name = "orc-agent-a2a"
version = "0.1.0"
description = "Orc Agent (A2A-compliant)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-server @ file:///Users/chrism1/Code/ai/agentic/a2a/a2a-server",
    "deprecated>=1.2.18",
    "google-adk>=0.2.0",
    "litellm>=1.67.0",
    "python-dotenv>=1.1.0",
]
